refractored the main file
data and functions are now in different files.
I added comments in the main file for better code readebility
