+++ data.py
@@ -1,0 +1,25 @@
+tasks = [
+    {"id":1 ,"name": "1000 pushups", "priority": 8},
+    {"id":2 ,"name": "Heel veel aardbeien eten", "priority": 3},
+    {"id":3 ,"name": "afvallen", "priority": 10},
+]
+decisionMakingQuips = [
+    {"text": "Thank you for your Critical Decision Making", "retry": False},
+    {"text": "You thought long and hard about that decision", "retry": False},
+    {"text": "Are you sure about that?", "retry": False},
+    {"text": "Hmmmm, interesting choice", "retry": False},
+    {"text": "You must have given that a lot of thought", "retry": False},
+    {"text": "I wouldn't have made that choice", "retry": False},
+    {"text": "NOOO PLEASE NOT THAT ONE", "retry": True},
+    {"text": "Sorry for the inconvenience, but that choice is not available at the moment", "retry": True},
+    {"text": "ERROR - Something went wrong", "retry": True}
+]
+
+processingQuips = [
+    {"text": "Processing...", "timer": 10},
+    {"text": "Lemme think about that one...", "timer": 20},
+    {"text": "Hhmmm, this won't take long...", "timer": 100},
+    {"text": "I'm thinking, one moment please...", "timer": 18},
+    {"text": "Thinking...", "timer": 20},
+    {"text": "Phoe Tough one", "timer": 500}
+]
